InputCommon: Split Motor from Touchscreen class
