+++ Source/Core/InputCommon/ControllerInterface/Touch/Touchscreen.h
@@ -39,19 +39,6 @@ private:
     const ButtonManager::ButtonType m_index;
     const float m_neg;
   };
-  class Motor : public Core::Device::Output
-  {
-  public:
-    Motor(int pad_id, ButtonManager::ButtonType index) : m_pad_id(pad_id), m_index(index) {}
-    ~Motor();
-    std::string GetName() const override;
-    void SetState(ControlState state) override;
-
-  private:
-    const int m_pad_id;
-    const ButtonManager::ButtonType m_index;
-    static void Rumble(int pad_id, double state);
-  };
 
 public:
   Touchscreen(int pad_id, bool accelerometer_enabled, bool gyroscope_enabled);
